Remove async eventemitter
This module was casuing bugs that prevented my MetaMask from working.

It kept throwing the error documented here:
https://github.com/MetaMask/metamask-extension/issues/3262

Tracking down the nature of the problem was difficult, but it seemed that async-eventemitter was over-iterating its listener array.

Trying to even find that repo was difficult, no repo link on npm, hadn't been updated in 5 months, I'm not sure why we chose to rely on this module, but it smells bad to me.

On removing it, my MetaMask build works fine again. Please don't merge without answering this question:
Why did we want an async event emitter? What response needed to be executed in series?