Migrate from CircleCI to GitHub Actions (#88)
GitHub Actions is now used for CI instead of CircleCI. The workflow
from the MetaMask module template has been used here, with the non-
applicable steps removed (changelog validation and allow-scripts).

The `test` npm script has been updated to no longer include the build
step. They now need to be run separately, as is the convention in most
of our repositories.

So far only Node.js v12 is used in CI, because there are dependency
incompatibilities with v14 and v16. Support will be added in a later PR
for v14 and v16.