Add `destroy` method to block tracker classes
PollingBlockTracker and SubscribeBlockTracker contain asynchronous code
that is run somewhat out-of-band. Because the NetworkController uses a
block tracker under the hood, we need to ensure that the block tracker
and its asynchronous code is properly stopped after each test that we
write, otherwise we may get warnings from Jest that promises are still
outstanding before the test run ends. Adding a `destroy` method makes
that possible.