Remove `BaseBlockTracker`
The base block tracker has been inlined into the polling block tracker
and the subscribe block tracker. The base class has been removed. This
was done to improve readability and allow for further simplications in
later PRs.

A `BlockTracker` type has been added to represent the common block
tracker interface. This type can be used anywhere we might want to use
a block tracker where we don't care which specific type it is.