PollingBlockTracker: Stop wrapping errors
If PollingBlockTracker fails to make a request, the error is wrapped
with a marker a more descriptive message (including the stack trace)
before being emitted under the "error" event or logged. This wrapping
may have been added in the past to assist with debugging, but it is not
necessary now, and it complicates the tests.