fix: Prevent delayed `getLatestBlock` responses
The recent `useCache` parameter for `getLatestBlock` let the caller
ask for a new block each time, which we thought would be an effective
way to ensure the block tracker never gets stuck with a stale block
indefinitely when it's not running.

However, this approach results in significant delays. Every call to
`getLatestBlock({ useCache: false })` waits for the _next_ block, even
if we have an extremely recent one. This isn't what we intended.

Instead this PR proposes that we use a deferred Promise to represent
a "pause" between requests, and _always_ return the cached response if
it's still pending, fetching otherwise. This ensures that we always
fetch if stale, but we return immediately if we have a "recent enough"
answer. This same Promise is initialized after the fetch call to
prevent too-frequent calls as well.

This should accomplish what we intended to achieve with the `useCache`
parameter, except without any downside. And it's more similar to the
previous implementation (it only differs in that it that it never
returns an extremely-stale response), so there is less concern about
this being a disruptive change.