Remove retry
The block tracker has a "retry" mechanism that will make the block
tracker run more rapidly (at 10x the normal frequency) after an error.
This is redundant alongside the new retry mechanisms we have in place
in the RPC service used by the underlying fetch middleware.

This retry has been removed. This simplifies the block tracker in a way
that may help with another problem we're currently investigating.