test: Prevent tests from inserting code synchronously after `setTimeout`
The timeout recorder utility used in unit tests could accidentally
allow a test to insert synchronous code after a `setTimeout` was
resolved. The methods to allow waiting for a pending `setTimeout` to
execute didn't clear pending Promises, so the code following those
calls in a test would run _immediately_ after the timeout was resolved,
which normally would not be possible for an external caller.

The code-under-test may (correctly) assume that a caller cannot run
code after a `setTimeout` resolves before it can. Tests that violate
this assumption unintentionally serve no practical purpose, and may be
confusing for contributors or require complex workarounds. This exact
problem cropped up in another draft PR currently open, #344.

To prevent this, pending Promises are resolved before returning when
waiting for a `setTimeout` to resolve.